change type of operating resources

+++ model/item/kind_firearm.go
@@ -23,7 +23,7 @@ type Firearm struct {
 	FoldRectractable   bool     `json:"foldRectractable" bson:"foldRectractable"`
 	RecoilVertical     int64    `json:"recoilVertical" bson:"recoilVertical"`
 	RecoilHorizontal   int64    `json:"recoilHorizontal" bson:"recoilHorizontal"`
-	OperatingResources int64    `json:"operatingResources" bson:"operatingResources"`
+	OperatingResources float64  `json:"operatingResources" bson:"operatingResources"`
 	DurabilityRatio    float64  `json:"durabilityRatio" bson:"durabilityRatio"`
 	HeatFactor         float64  `json:"heatFactor" bson:"heatFactor"`
 	Slots              Slots    `json:"slots" bson:"slots"`
